add db inspect
